+++ ext/DeviceUtilsRecursiveArrayToolsExt.jl
@@ -14,8 +14,11 @@ function Adapt.adapt_structure(to::AbstractDevice, x::DiffEqArray)
     return DiffEqArray(map(Base.Fix1(adapt, to), x.u), x.t)
 end
 
-function DeviceUtils.get_device(x::Union{VectorOfArray, DiffEqArray})
-    return mapreduce(DeviceUtils.get_device, DeviceUtils.__combine_devices, x.u)
+for op in (:_get_device, :_get_device_type)
+    @eval function DeviceUtils.$op(x::Union{VectorOfArray, DiffEqArray})
+        length(x.u) == 0 && return $(op == :_get_device ? nothing : Nothing)
+        return mapreduce(DeviceUtils.$op, DeviceUtils.__combine_devices, x.u)
+    end
 end
 
 end
