+++ ext/DeviceUtilsReverseDiffExt.jl
@@ -3,11 +3,15 @@ module DeviceUtilsReverseDiffExt
 using DeviceUtils: DeviceUtils
 using ReverseDiff: ReverseDiff
 
-@inline function DeviceUtils.get_device(x::ReverseDiff.TrackedArray)
-    return DeviceUtils.get_device(ReverseDiff.value(x))
-end
-@inline function DeviceUtils.get_device(x::AbstractArray{<:ReverseDiff.TrackedReal})
-    return DeviceUtils.get_device(ReverseDiff.value.(x))
+for op in (:_get_device, :_get_device_type)
+    @eval begin
+        function DeviceUtils.$op(x::ReverseDiff.TrackedArray)
+            return DeviceUtils.$op(ReverseDiff.value(x))
+        end
+        function DeviceUtils.$op(x::AbstractArray{<:ReverseDiff.TrackedReal})
+            return DeviceUtils.$op(ReverseDiff.value.(x))
+        end
+    end
 end
 
 end
