+++ ext/DeviceUtilsTrackerExt.jl
@@ -5,14 +5,16 @@ using DeviceUtils: DeviceUtils, AMDGPUDevice, CUDADevice, MetalDevice,
                       oneAPIDevice
 using Tracker: Tracker
 
-@inline function DeviceUtils.get_device(x::Tracker.TrackedArray)
-    return DeviceUtils.get_device(Tracker.data(x))
-end
-@inline function DeviceUtils.get_device(x::AbstractArray{<:Tracker.TrackedReal})
-    return DeviceUtils.get_device(Tracker.data.(x))
+for op in (:_get_device, :_get_device_type)
+    @eval begin
+        DeviceUtils.$op(x::Tracker.TrackedArray) = DeviceUtils.$op(Tracker.data(x))
+        function DeviceUtils.$op(x::AbstractArray{<:Tracker.TrackedReal})
+            return DeviceUtils.$op(Tracker.data.(x))
+        end
+    end
 end
 
-@inline DeviceUtils.__special_aos(::AbstractArray{<:Tracker.TrackedReal}) = true
+DeviceUtils.__special_aos(::AbstractArray{<:Tracker.TrackedReal}) = true
 
 for T in (AMDGPUDevice, AMDGPUDevice{Nothing}, CUDADevice,
     CUDADevice{Nothing}, MetalDevice, oneAPIDevice)
